Added comments and removed unnecessary files
